feat(docs.ws): `Next.js` successfully generates all data feeds pages
Implement `Data Feeds > feed > [feed]` pages as a `statically prerendered RSC`
